[jk] Move retry_pipeline_run method to PipelineRun model
